requirements.txt
Flask==1.1.0 
pytest==5.0.1
Flask-WTF==0.14.2 
python-dotenv==0.10.3
Flask-SQLAlchemy==2.4.0